logic/requirements.txt
Flask==2.2.5
pytesseract==0.3.10
opencv-python==4.8.0.74
Pillow==9.5.0
gTTS==2.4.6
python-dotenv==1.0.0
